- Move painting functions into PaintUtil. - Refactor.
